+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- fix(remix): Change `handleError` in server entry (#621)
+
 ## 3.25.0
 
 - feat(react-native): Add support for Expo projects (#505)